backend/requirements.txt
applicationinsights==0.11.7
colour==0.1.5
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.13
django-cors-headers==3.1.1
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.10
djangorestframework-csv==2.1.0
djangorestframework-filters==0.10.2
drf-yasg==1.17.1
environs==4.1.0
djangorestframework-xml==1.4.0
gunicorn==20.1.0
furl==2.0.0
pyexcel==0.5.14
seqeval==0.0.6
whitenoise[brotli]==4.1.2
conllu==1.3.2
numpy==1.14.1